# CMake: improve testsuite driver performance