Some preliminary results:
```
% ctest -R "petsc_step-27.mpirun=2.debug" # first run:
1/2 Test #9807: test_dependency/mpi.petsc_step-27.debug.executable ...   Passed   14.15 sec
2/2 Test #9811: mpi/petsc_step-27.mpirun=2.debug .....................   Passed    3.08 sec
% ctest -R "petsc_step-27.mpirun=2.debug" # rerun:
1/2 Test #9807: test_dependency/mpi.petsc_step-27.debug.executable ...   Passed    0.15 sec
2/2 Test #9811: mpi/petsc_step-27.mpirun=2.debug .....................   Passed    0.14 sec
```
As compared to a baseline test with make:
```
% ctest -R "petsc_step-27.mpirun=2.debug" # first run
1/2 Test #9807: test_dependency/mpi.petsc_step-27.debug.executable ...   Passed   16.38 sec
2/2 Test #9811: mpi/petsc_step-27.mpirun=2.debug .....................   Passed    5.42 sec
% ctest -R "petsc_step-27.mpirun=2.debug" # rerun
1/2 Test #9807: test_dependency/mpi.petsc_step-27.debug.executable ...   Passed    2.36 sec
2/2 Test #9811: mpi/petsc_step-27.mpirun=2.debug .....................   Passed    2.33 sec
```