@tamiko 2023-06-26
This is a very preliminary version of some more massive rewriting effort. My goal is to significantly reduce the overhead introduced by `make` in the testsuite.